pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gtd-todos"
version = "0.1.0"
description = "Local-first GTD TODO application built with FastAPI, HTMX-ready templates, and SQLite."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "bleach>=6.2,<7.0",
  "fastapi>=0.115,<1.0",
  "jinja2>=3.1,<4.0",
  "markdown-it-py>=3.0,<4.0",
  "python-multipart>=0.0.20,<1.0",
  "sqlmodel>=0.0.22,<0.0.23",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "mypy>=1.15,<2.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
  "ruff>=0.11,<0.12",
  "types-bleach>=6.2.0.20250809,<7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["templates/*.html", "static/*.css"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
pretty = true
